+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/types/TypeSystem.java
@@ -183,7 +183,9 @@ public final class TypeSystem {
      * @param bootstrapResourceLoader Classloader used to resolve class files
      *                                to populate the fields of the new type
      *                                system
+     * @deprecated Since 7.27.0. Use {@link #usingClasspath(Classpath)} instead.
      */
+    @Deprecated
     public static TypeSystem usingClassLoaderClasspath(ClassLoader bootstrapResourceLoader) {
         return usingClasspath(Classpath.forClassLoader(bootstrapResourceLoader));
     }
@@ -313,7 +315,7 @@ public final class TypeSystem {
     private JClassSymbol getBootStrapSymbol(Class<?> clazz) {
         AssertionUtil.requireParamNotNull("clazz", clazz);
         JClassSymbol sym = resolver.resolveClassFromBinaryName(clazz.getName());
-        return Objects.requireNonNull(sym, "sym");
+        return Objects.requireNonNull(sym, "symbol for " + clazz + " was null");
     }
 
     private @NonNull JPrimitiveType createPrimitive(PrimitiveTypeKind kind, Class<?> box) {